requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord-interactions==0.4.0
discord-py-interactions==5.11.0
discord-py-slash-command==4.2.1
discord-typings==0.7.0
discord.py==2.4.0
emoji==2.9.0
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.0.4
outcome==1.3.0.post0
packaging==23.2
pycparser==2.21
PyNaCl==1.5.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
selenium==4.23.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tomli==2.0.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
webdriver-manager==4.0.1
websocket-client==1.8.0
Werkzeug==3.1.1
wsproto==1.2.0
yarl==1.9.4